backend/requirements.txt
Django==5.2.3
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
pillow==11.2.1
psycopg2==2.9.10
